+++ js/util.js
@@ -1,0 +1,6 @@
+// eslint-disable-next-line no-unused-vars
+function isLessThanMax(string, maxLength) {
+  return string.length <= maxLength;
+}
+
+export {isLessThanMax};
